web server can be built in VS2005; server compiles for 64bit; minor fixes

+++ src/server/core/audit.cpp
@@ -76,7 +76,7 @@ void NXCORE_EXPORTABLE WriteAuditLog(const TCHAR *pszSubsys, BOOL bSuccess, DWOR
 {
 	TCHAR *pszQuery, *pszText, *pszEscText;
 	va_list args;
-	int nBufSize;
+	size_t nBufSize;
 	CSCPMessage msg;
 
 	nBufSize = _tcslen(pszFormat) + NumChars(pszFormat, _T('%')) * 256 + 1;
